+++ apps/rhf-mui-demo/src/forms/autocomplete/index.tsx
@@ -8,7 +8,6 @@ import Typography from '@mui/material/Typography';
 import RHFCountrySelect, { countryList, CountryISO } from '@nish1896/rhf-mui-components/mui/country-select';
 import RHFAutocomplete from '@nish1896/rhf-mui-components/mui/autocomplete';
 import RHFMultiAutocomplete from '@nish1896/rhf-mui-components/mui/multi-autocomplete';
-import { toast } from 'react-toastify';
 import {
   FormContainer,
   FormState,
@@ -18,7 +17,7 @@ import {
 } from '@/components';
 import { Colors } from '@/types';
 import { IPLTeams } from '@/constants';
-import { printJSONObject, showToastMessage } from '@/utils';
+import { showToastMessage } from '@/utils';
 
 type AirportInfo = {
   iataCode: string;
@@ -65,7 +64,7 @@ const AutocompleteForm = () => {
   const filteredCountries = countryList.filter(country => country.name.length > 5);
 
   function onFormSubmit(formValues: FormSchema) {
-    showToastMessage(printJSONObject(formValues));
+    showToastMessage(formValues);
   }
 
   return (
@@ -141,6 +140,8 @@ const AutocompleteForm = () => {
                 }
               }}
               getLimitTagsText={more => `+${more} Color(s)`}
+              helperText="Select at least 2 colors"
+              formControlLabelProps={{ sx: { color: 'royalblue' } }}
               errorMessage={errors?.colors?.message}
             />
           </Grid>