+++ apps/rhf-mui-demo/src/forms/checkbox-and-radiogroup-with-zod/index.tsx
@@ -15,7 +15,7 @@ import {
 } from '@/components';
 import { CountriesList } from '@/constants';
 import { Gender } from '@/types';
-import { printJSONObject } from '@/utils';
+import { showToastMessage } from '@/utils';
 import { formSchema, PersonInfo } from './validation';
 
 const CheckboxRadioZodForm = () => {
@@ -29,7 +29,7 @@ const CheckboxRadioZodForm = () => {
   });
 
   function onFormSubmit(formValues: PersonInfo) {
-    alert(`Form Submitted with values: \n ${printJSONObject(formValues)}`);
+    showToastMessage(formValues);
   }
 
   return (