+++ apps/rhf-mui-demo/src/forms/complete-form/index.tsx
@@ -37,7 +37,7 @@ import {
 import { CountriesList, IPLTeams, GroceryList, HobbiesList } from '@/constants';
 import { useThemeContext } from '@/theme';
 import { Colors, Gender, Sports, Person } from '@/types';
-import { printJSONObject } from '@/utils';
+import { showToastMessage } from '@/utils';
 
 type FormSchema = Person & { disableAllFields?: boolean; }
 
@@ -71,7 +71,7 @@ const CompleteForm = () => {
   }
 
   function onFormSubmit(formValues: FormSchema) {
-    alert(`Form Submitted with values: \n ${printJSONObject(formValues)}`);
+    showToastMessage(formValues);
   }
 
   return (