+++ apps/rhf-mui-demo/src/forms/select-with-class-validator/index.tsx
@@ -16,7 +16,7 @@ import {
 } from '@/components';
 import { IPLTeams, Currencies } from '@/constants';
 import { Colors } from '@/types';
-import { printJSONObject } from '@/utils';
+import { showToastMessage } from '@/utils';
 
 const randomNumbers = [23, 56, 67, 32, 68, 54, 90];
 
@@ -32,7 +32,7 @@ const SelectFormWithClassValidator = () => {
   });
 
   function onFormSubmit(formValues: FormSchema) {
-    alert(`Form Submitted with values: \n ${printJSONObject(formValues)}`);
+    showToastMessage(formValues);
   }
 
   return (