+++ apps/rhf-mui-demo/src/forms/switch-slider-rating-with-superstruct/index.tsx
@@ -16,7 +16,7 @@ import {
   FieldVariantInfo,
   SubmitButton
 } from '@/components';
-import { printJSONObject } from '@/utils';
+import { showToastMessage } from '@/utils';
 
 const orangeTheme = createTheme({
   palette: {
@@ -41,7 +41,7 @@ const SwitchSliderRatingFormWithSuperstruct = () => {
   });
 
   function onFormSubmit(formValues: FormSchema) {
-    alert(`Form Submitted with values: \n ${printJSONObject(formValues)}`);
+    showToastMessage(formValues);
   }
 
   return (