+++ apps/rhf-mui-demo/src/utils/messages.ts
@@ -14,6 +14,6 @@ export function pageDescriptionMsg(fieldName: string) {
   return `Input props and usage of ${fieldName}`;
 }
 
-export function printJSONObject(obj: object) {
-  return JSON.stringify(obj, null, 4);
+export function printJSONObject(obj: object | string) {
+  return JSON.stringify(obj, null, 2);
 }