+++ apps/rhf-mui-demo/src/utils/toast.ts
@@ -1,10 +1,12 @@
 import { toast } from 'react-toastify';
 import { ToastDialog } from '@/components';
+import { printJSONObject } from '@/utils';
 
-export function showToastMessage(msg: string) {
-	toast.success(ToastDialog, {
+export function showToastMessage(msg: object | string) {
+	const msgContent = typeof msg === 'string' ? msg : printJSONObject(msg);
+	toast(ToastDialog, {
 		data: {
-			content: msg,
+			content: msgContent
 		}
 	});
 }